tensorflow_impl/requirements.txt
tensorflow
numpy==1.18
grpcio
grpcio-tools
tensorflow_datasets